% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{add_miss_ap3_cols}
\alias{add_miss_ap3_cols}
\title{Check a data.frame for columns required to calculate Apache III score}
\usage{
add_miss_ap3_cols(df)
}
\arguments{
\item{df}{A data frame with (case insensitive) column names conforming to the
ANZICS APD Data Dictionary Apache III subset of fields}
}
\value{
The same data frame that was passed as an argument with all required
columns added
}
\description{
Given a \code{data.frame} or \code{tibble} this function checks for the minimum
required columns to calculate an Apache III score. The fields are documented
in the ANZICS Data Dictionary. An example of the minimum required data set is
available \link{ap3_appendix_example}. All missing fields are added with a column
containing entirely \code{NA_real_} and a message is printed to the console
describing which columns were added.
}
\details{
add_miss_ap3_cols
}
\examples{
# a complete data.frame
data("ap3_appendix_example")
add_miss_ap3_cols(ap3_appendix_example)
# a data.frame with many required columns missing
sparse.data <- data.frame(hrhi=120,maphi=55)
add_miss_ap3_cols(sparse.data)
}
\references{
https://www.anzics.org/wp-content/uploads/2021/03/ANZICS-APD-Dictionary-Version-6.1.pdf (ANZICS APD Data Dictionary)
}
\seealso{
Other Apache3: 
\code{\link{ap3_appendix_example}},
\code{\link{ap3_comorbidity_scores}},
\code{\link{apache3}()},
\code{\link{apache3_comorb}()},
\code{\link{apache3_resp}()},
\code{\link{lab_conversions}}
}
\concept{Apache3}
