% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap3_appendix_example.R
\docType{data}
\name{ap3_appendix_example}
\alias{ap3_appendix_example}
\title{Apache III Example ICU Patient Data}
\format{
\subsection{\code{ap3_appendix_example}}{

A data frame with 2 rows and 22 columns:
\describe{
\item{IcuEpisodeId}{A unique identified for each ICU admission episode}
\item{age}{Age in years}
\item{comorb}{Apache III Comorbidity diagnosis as a string - use \code{data("ap3comorbidity_scores")} to see acceptable terms & associated score}
\item{elect_surg}{Was the ICU admission after elective surgery (comorb is not assigned a score)}
\item{hr}{Heart Rate in beats per minute}
\item{map}{Mean Arterial Blood Pressure in mmHg}
\item{temp}{Temperature in °C}
\item{rr}{Respirator Rate in breaths per minute. If ventilated, it is the total respiratory rate (mandatory + spontaneous respirations)}
\item{rr_vent}{Was the patient on a ventilator when the respiratory rate was recorded}
\item{hct}{Haematocrit (as a \% or fraction)}
\item{wcc}{White Cell Count × 10^9/L}
\item{creat}{Creatinine in µmol/L}
\item{urineop}{Urine Output in mL per 24 hours}
\item{urea}{Urea in mmol/L}
\item{Na}{Sodium in mmol/L}
\item{album}{Albumin in g/L}
\item{bili}{Bilirubin in µmol/L}
\item{gluc}{Blood Glucose in mmol/L}
\item{gcseye}{The Eyes component of the Glasgow Coma Scale (GCS) - 1 to 4}
\item{gcsverb}{The Verbal component of the GCS - 1 to 5}
\item{gcsmotor}{The Motor Component of the GCS - 1 to 6}
\item{gcs_sedated}{Was a meaningful GCS impossible to obtain due to medical sedation}
\item{arf}{Was the patient in Acute Renal Failure within 24 hours of admission, defined as 1) Urine < 410 ml in 1st 24hr 2) Creatinine >133 μmol/L 3) No chronic dialysis}
}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/1959406/}
}
\usage{
data(ap3_appendix_example)
}
\description{
Data representing the variables for the example patients described in
appendix B of the original paper reporting the Apache III score.
}
\details{
The data format is according to the ANZICS APD data dictionary
}
\seealso{
Other Apache3: 
\code{\link{add_miss_ap3_cols}()},
\code{\link{ap3_comorbidity_scores}},
\code{\link{apache3}()},
\code{\link{apache3_comorb}()},
\code{\link{apache3_resp}()},
\code{\link{lab_conversions}}
}
\concept{Apache3}
\keyword{datasets}
