% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap3_appendix_example.R
\docType{data}
\name{ap3_appendix_gas}
\alias{ap3_appendix_gas}
\title{Apache III Example ICU ABG Data}
\format{
\subsection{\code{ap3_appendix_gas}}{

A data frame with 2 rows and 22 columns:
\describe{
\item{IcuEpisodeId}{A unique identified for each ICU admission episode}
\item{intubated}{Was the patient intubated at the time the gas was obtained}
\item{pao2}{Partial pressure of Oxygen in the arterial gas in mmHg}
\item{fio2}{Fraction of inspired oxygen the patient was receiving at the time the gas was obtained}
\item{ph}{pH of the sample}
\item{paco2}{Partial pressure of Carbon Dioxide in the arterial gas in mmHg}
}
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/1959406/}
}
\usage{
data(ap3_appendix_gas)
}
\description{
Data representing arterial blood gasses obtained in the first 24 hours of
ICU admission - taken from the original paper reporting the Apache III score
}
\details{
The data format is similar to the ANZICS APD data dictionary, although PaO2
is used rather than PO2 (and PaCO2) The \code{apache3_resp} functions in this
package work with either abbreviation, but  PaO2 is clearer when
differentiating arterial from venous, central venous or mixed venous gasses.
}
\concept{Apache}
\keyword{datasets}
