% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ap3_comorbidity_scores.R
\docType{data}
\name{ap3_comorbidity_scores}
\alias{ap3_comorbidity_scores}
\title{Conversion factors to SI units for lab data}
\format{
\subsection{\code{ap3_comorbidity_scores}}{

A named integer vector with labels representing the names of the comorbidities
and values being the Apache III score to be assigned if that comorbidity is present
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/1959406/}
}
\usage{
data(ap3_comorbidity_scores)
}
\description{
The row.names represent the lab test and the other columns provide information to convert
the test result to the required units (if the field data is in alternative units)
}
\seealso{
Other Apache3: 
\code{\link{ap3_appendix_example}},
\code{\link{ap3_appendix_gas}},
\code{\link{lab_conversions}}
}
\concept{Apache3}
