% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache3.R
\name{apache3}
\alias{apache3}
\title{Generate Apache III composite scores}
\usage{
apache3(df)
}
\arguments{
\item{df}{a \code{data.frame} or \code{tibble}  which complies with nomenclature outlined in the
ANZICS data dictionary. see \code{\link[=add_miss_ap3_cols]{add_miss_ap3_cols()}} and example data from the article
\link{ap3_appendix_example}.}
}
\value{
A \code{tibble} containing age, physiology and lab components of the
APACHE III score for each \code{IcuEpisodeId}
}
\description{
Given a \code{data.frame} or tibble  which complies with nomenclature outlined in the
ANZICS data dictionary, this will generate a data-set containing each component of the
APACHE III score for each \code{IcuEpisodeId}
}
\details{
See the \href{../doc/apache3.html}{\code{vignette("Apache3", package = "icuscorer")}}
vignette
}
\examples{
data("ap3_appendix_example")
ap3.scores <- apache3(ap3_appendix_example)
ap3.scores
}
\references{
Knaus WA, Wagner DP, Draper EA, et al. The APACHE III prognostic system. Risk prediction of hospital mortality for critically ill hospitalized adults. Chest. 1991 Dec;100(6):1619-36. doi: 10.1378/chest.100.6.1619. PMID: 1959406.

https://www.anzics.org/wp-content/uploads/2021/03/ANZICS-APD-Dictionary-Version-6.1.pdf (ANZICS APD Data Dictionary)
}
\seealso{
Other apache3: 
\code{\link{add_miss_ap3_cols}()},
\code{\link{apache3_comorb}()},
\code{\link{apache3_resp}()}
}
\concept{apache3}
