% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache3_comorb.R
\name{apache3_comorb}
\alias{apache3_comorb}
\title{Generate Apache III scores from comorbidities}
\usage{
apache3_comorb(comorbs)
}
\arguments{
\item{comorbs}{A \code{data.frame} or tibble  which contains the columns (case
insensitive) \code{IcuEpisodeId}, \code{comorb} (character) and
\code{elect_surg} (logical).}
}
\value{
A \code{tibble} containing the comorbidity score for each
\code{IcuEpisodeId}. Comorbidity scoring details are available in these data
\link{ap3_comorbidity_scores}
}
\description{
Given a \code{data.frame} or tibble  of co-morbidities containing a character
column of diagnoses, selects the highest co-morbidity per \code{IcuEpisodeId}
or 0 if the ICU admission was following elective surgery
}
\details{
This is unlikely to be used in production given various diagnostic
categorisations and ICU registries in use. This function is therefore
provided primarily for completeness and for use in the examples.
}
\examples{
data("ap3_appendix_example")
ap3.comorb.score <- apache3_comorb(ap3_appendix_example)
ap3.comorb.score
}
\references{
Knaus WA, Wagner DP, Draper EA, et al. The APACHE III prognostic system. Risk prediction of hospital mortality for critically ill hospitalized adults. Chest. 1991 Dec;100(6):1619-36. doi: 10.1378/chest.100.6.1619. PMID: 1959406.
}
\seealso{
Other Apache3: 
\code{\link{add_miss_ap3_cols}()},
\code{\link{ap3_appendix_example}},
\code{\link{ap3_comorbidity_scores}},
\code{\link{apache3}()},
\code{\link{apache3_resp}()},
\code{\link{lab_conversions}}
}
\concept{Apache3}
