% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache3_resp.R
\name{apache3_resp}
\alias{apache3_resp}
\title{Generate Apache III scores from Arterial Blood Gasses}
\usage{
apache3_resp(abgs)
}
\arguments{
\item{abgs}{A \code{data.frame} or \code{tibble}  which contains the columns (case
insensitive) \code{IcuEpisodeId}, \code{pH}, \code{PaCO2} or \code{PCO2}, \code{PaO2} or \code{PO2}, and
\code{intubated} (logical). Example gasses from the original article appendix are
available \link{ap3_appendix_gas}.}
}
\value{
A \code{tibble} containing the oxygen (O2) and pH/CO2 score for each
\code{IcuEpisodeId}
}
\description{
Given a \code{data.frame} or tibble  of arterial blood gas results, this will
generate a data-set containing the O2 and pH/CO2 component of the APACHE III
score for each \code{IcuEpisodeId}
}
\examples{
data("ap3_appendix_gas")
ap3.gas.scores <- apache3_resp(ap3_appendix_gas)
ap3.gas.scores
}
\references{
Knaus WA, Wagner DP, Draper EA, et al. The APACHE III prognostic system. Risk prediction of hospital mortality for critically ill hospitalized adults. Chest. 1991 Dec;100(6):1619-36. doi: 10.1378/chest.100.6.1619. PMID: 1959406.
}
\seealso{
Other apache3: 
\code{\link{add_miss_ap3_cols}()},
\code{\link{apache3}()},
\code{\link{apache3_comorb}()}
}
\concept{apache3}
