% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assign_scores}
\alias{assign_scores}
\title{Assign scores to a vector}
\usage{
assign_scores(x, delimiters, score_assigned, rightmost_closed = FALSE)
}
\arguments{
\item{x}{A vector of values which should be assigned the relevant score, according to the partitions defined by 'delimiters'}

\item{delimiters}{A vector containing the partition boundaries}

\item{score_assigned}{The score associated with the partitions provided with the
delimiters parameter. The first argument is assigned if the value of x is less that
the value of the first delimiter, and the last value assigned if x is greater than the
final delimiter. The length of the score.assigned vector should be 1 element greater than the delimiters vector}

\item{rightmost_closed}{whether the delimiters are right closed. The default is false
meaning if the value is equal to the first delimiter, the 2nd score is assigned
and if equal to the last delimiter, the final score is assigned}
}
\value{
A vector of the same length as x, scored according to arguments provided
}
\description{
Assign a scoring system across a vector based on arbitrary cut-points and scores corresponding to each of the intervals
}
\details{
assign_scores
}
\examples{
assign_scores(1:20, c(3, 11, 14), c(0, 4, 7, 9))
}
\seealso{
Other clinical_scoring_utilities: 
\code{\link{if_else_fns}()},
\code{\link{replace_matches}()},
\code{\link{vector_to_logical}()}
}
\concept{clinical_scoring_utilities}
