% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_conversions.R
\name{F_to_C}
\alias{F_to_C}
\title{Fahrenheit conversion to Centigrade}
\usage{
F_to_C(f_temp)
}
\arguments{
\item{f_temp}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Convert degrees Fahrenheit temperatures to degrees Celsius.
While this function is a ubiquitous teaching example of writing a function
within an R package, U.S.based users of this package may well require it
for this library if 'TEMPHI" and 'TEMPLO' columns are in degrees Fahrenheit!
}
\examples{
temp1 <- F_to_C(50)
temp2 <- F_to_C(c(50, 63, 23))
}
