% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icuscorer-package.R
\docType{package}
\name{icuscorer-package}
\alias{icuscorer}
\alias{icuscorer-package}
\title{icuscorer: Intensive care patient risk Scoring systems}
\description{
Physiological, lab and diagnostic data (in a data.frame) are used to calculate published, validated risk scores for admissions to intensive care units (ICU). Risk scores are returned per ICU episode, as individual score components (pH score etc.). Currently only implemented for Acute Physiology And Chronic Health Evaluation (Apache) III score, but other ICU scoring systems will be added.
}
\author{
\strong{Maintainer}: McSharry Brent \email{bmcshaz@gmail.com} (\href{https://orcid.org/0009-0000-9735-6158}{ORCID})

}
\keyword{internal}
