% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{if_else_fns}
\alias{if_else_fns}
\title{vectorised if_else with conditional functions}
\usage{
if_else_fns(x, condition, if_true, if_false, .missing = NA_real_)
}
\arguments{
\item{x}{Any vector}

\item{condition}{A logical vector the same length as x corresponding to the if_else condition}

\item{if_true}{The function to be executed on all values of x where condition is TRUE. Must be a vectorised function
should return vector of the same length as the vector passed to it.}

\item{if_false}{As per the if_true argument, but where condition is FALSE.}

\item{.missing}{The value to be returned for elements where condition is NA. Default NA_real_}
}
\value{
A vector of the same length as x, with the returned data from the functions
executed acording to the condition vector
}
\description{
vectorised if_else which executes true and false functions passed a  vector
filtered on the condition
}
\details{
Similar to dplyr if_else, but accepts a function for the true and false arguments.
The true and false function will be executed once each, with the argument passed
to the function being a vector containing all data meeting the condition.
It expects a vector of the same length as that passed in the return value of the function
and this returned vector is then placed into a corresponding position in the final vector.

It could be considered a vectorised form of purrr::map_if

if_else_fns
}
\examples{
if_else_fns(1:20, 1:20 \%\% 2 == 0, \(x) {
  cat("passed arg len:", length(x), "\n")
  -x
}, \(x) x + 20)
}
\seealso{
Other clinical_scoring_utilities: 
\code{\link{assign_scores}()},
\code{\link{replace_matches}()},
\code{\link{vector_to_logical}()}
}
\concept{clinical_scoring_utilities}
