% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lab_conversions.R
\docType{data}
\name{lab_conversions}
\alias{lab_conversions}
\title{Conversion factors to SI units for lab data}
\format{
\subsection{\code{lab_conversions}}{

A data frame with 7 rows and 3 columns:
\describe{
\item{row.names}{The variable name as per the ANZICS data dictionary}
\item{reqired_units}{The units of measurment this library assumes the field has been converted to}
\item{alternative_units}{The other units of measurement commonly in use, to which the conversion factor applies}
\item{to_required}{The multiplier to convert alternative_units to required_units}
}
}
}
\usage{
data(lab_conversions)
}
\description{
The row.names represent the lab test and the other columns provide information to convert
the test result to the required units (if the field data is in alternative units)
}
\examples{
data.frame(
  creat = c(2.2, 1.8) * lab_conversions["CREAT", "to_required"],
  urea = c(85, 22) * lab_conversions["UREA", "to_required"],
  bili = c(3.3, 2.5) * lab_conversions["BILI", "to_required"],
  gluc = c(246, 190) * lab_conversions["GLUC", "to_required"]
)
}
\seealso{
Other Apache3: 
\code{\link{add_miss_ap3_cols}()},
\code{\link{ap3_appendix_example}},
\code{\link{ap3_comorbidity_scores}},
\code{\link{apache3}()},
\code{\link{apache3_comorb}()},
\code{\link{apache3_resp}()}
}
\concept{Apache3}
