% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apache3_gcs.R
\name{multilookup_of_array}
\alias{multilookup_of_array}
\title{Look up multiple elements in a mult-dimensional array}
\usage{
multilookup_of_array(lookup_array, findvalues_matrix)
}
\arguments{
\item{lookup_array}{The array to be searched}

\item{findvalues_matrix}{a matrix with 1 column for each dimension of 'findvalues'.
The values in each row represent the indexes corresponding to each dimension.}
}
\value{
A vector of elements - one element for each row of 'findvalues_matrix'
}
\description{
Look up multiple elements in a mult-dimensional array
}
\examples{
l <- array(1:60, dim=5:3)
l
f <- matrix(c(4, 2, 1, 3, 4, 2, 4, 2, 1), ncol=3)
f
multilookup_of_array(l, f)

}
