% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{replace_matches}
\alias{replace_matches}
\title{Replace matches}
\usage{
replace_matches(x, dictionary, unmatched_as_na = TRUE)
}
\arguments{
\item{x}{A character vector of values to search and replace. Matches in \code{x} will ignore both case and leading
or trailing white-space}

\item{dictionary}{A named vector or a list with the names being the strings searched for and the
values being the values to replace occurrences with}

\item{unmatched_as_na}{when \code{x} contains a string not found in \code{dictionary}, return NA (default).
If FALSE the corresponding (unmatched) value of \code{x} is returned}
}
\value{
A vector of the same length as \code{x}, with replacements made where matches are found
}
\description{
Replace exact matches with the corresponding value provided

replace_matches
}
\examples{
replace_matches(letters[1:10], list(a = 3, c = 7, f = 12))
# alternatively using a named vector
replace_matches(letters[1:10], c(a = 3, c = 7, f = 12))
replace_matches(letters[1:10], c(a = 3, c = 7, f = 12), unmatched_as_na = FALSE)
}
\seealso{
Other clinical_scoring_utilities: 
\code{\link{assign_scores}()},
\code{\link{if_else_fns}()},
\code{\link{vector_to_logical}()}
}
\concept{clinical_scoring_utilities}
